backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
scikit-learn
numpy